+++ Cercle.h
@@ -24,17 +24,17 @@ private:
 	//!
 	//! rayon du cercle
 	//!
-	int rayon;
+	double rayon;
 
 public:
 	//!
 	//! \brief Constructeur par point
 	//!
-	Cercle(const Couleur::couleurs &, const Point &, int r);
+	Cercle(const Couleur::couleurs &, const Point &, double r);
 	//!
 	//! \brief Constructeur par coordonnees
 	//!
-	Cercle(const Couleur::couleurs &, int x, int y, int r);
+	Cercle(const Couleur::couleurs &, double x, double y, double r);
 	//!
 	//! \brief Destructeur
 	//!
@@ -47,15 +47,15 @@ public:
 	//!
 	//! \brief override
 	//!
-	virtual FormeGeometrique* translation(int l, int h);
+	virtual FormeGeometrique* translation(double l, double h);
 	//!
 	//! \brief override
 	//!
-	virtual FormeGeometrique* homothetie(int x, int y, double coeff);
+	virtual FormeGeometrique* homothetie(double x, double y, double coeff);
 	//!
 	//! \brief override
 	//!
-	virtual FormeGeometrique* rotation(int x, int y, double angle);
+	virtual FormeGeometrique* rotation(double x, double y, double angle);
 
 	//!
 	//! \brief override
