+++ FormeCompliquee.h
@@ -55,7 +55,7 @@ public:
 	//! \param h : distance en y de la translation
 	//! \return l'image par la translation de la forme appelante 
 	//!
-	virtual FormeGeometrique* translation(int l, int h);
+	virtual FormeGeometrique* translation(double l, double h);
 	//!
 	//! \brief Homothetie de la forme
 	//! \param x : x du centre de homothetie
@@ -63,7 +63,7 @@ public:
 	//! \param coeff : coefficient de l'homothetie
 	//! \return l'image par l' homothetie de la forme appelante 
 	//!
-	virtual FormeGeometrique* homothetie(int x, int y, double coeff);
+	virtual FormeGeometrique* homothetie(double x, double y, double coeff);
 	//!
 	//! \brief Rotation de la forme
 	//! \param x : x du centre de rotation
@@ -71,7 +71,7 @@ public:
 	//! \param angle : angle de la rotation en degres
 	//! \return l'image par la rotation de la forme appelante 
 	//!
-	virtual FormeGeometrique* rotation(int x, int y, double angle);
+	virtual FormeGeometrique* rotation(double x, double y, double angle);
 
 	//!
 	//! \brief Caster l'objet en string
